magento-1.6.0.0

+++ app/code/core/Mage/Cms/Model/Wysiwyg/Images/Storage.php
@@ -20,7 +20,7 @@
  *
  * @category    Mage
  * @package     Mage_Cms
- * @copyright   Copyright (c) 2010 Magento Inc. (http://www.magentocommerce.com)
+ * @copyright   Copyright (c) 2011 Magento Inc. (http://www.magentocommerce.com)
  * @license     http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
@@ -59,12 +59,13 @@ class Mage_Cms_Model_Wysiwyg_Images_Storage extends Varien_Object
      */
     public function getDirsCollection($path)
     {
-        $subDirectories = Mage::getModel('core/file_storage_directory_database')->getSubdirectories($path);
-
-        foreach ($subDirectories as $directory) {
-            $fullPath = rtrim($path, DS) . DS . $directory['name'];
-            if (!file_exists($fullPath)) {
-                mkdir($fullPath, 0777, true);
+        if (Mage::helper('core/file_storage_database')->checkDbUsage()) {
+            $subDirectories = Mage::getModel('core/file_storage_directory_database')->getSubdirectories($path);
+            foreach ($subDirectories as $directory) {
+                $fullPath = rtrim($path, DS) . DS . $directory['name'];
+                  if (!file_exists($fullPath)) {
+                    mkdir($fullPath, 0777, true);
+                }
             }
         }
 
@@ -232,7 +233,9 @@ class Mage_Cms_Model_Wysiwyg_Images_Storage extends Varien_Object
 
         $io = new Varien_Io_File();
 
-        Mage::getModel('core/file_storage_directory_database')->deleteDirectory($path);
+        if (Mage::helper('core/file_storage_database')->checkDbUsage()) {
+            Mage::getModel('core/file_storage_directory_database')->deleteDirectory($path);
+        }
         if (!$io->rmdir($path, true)) {
             Mage::throwException(Mage::helper('cms')->__('Cannot delete directory %s.', $path));
         }
